+++ HeterogeneousCore/AlpakaCore/interface/CopyToDeviceCache.h
@@ -3,8 +3,8 @@
 
 #include <alpaka/alpaka.hpp>
 
-#include "HeterogeneousCore/AlpakaCore/interface/QueueCache.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/CopyToDevice.h"
+#include "HeterogeneousCore/AlpakaInterface/interface/QueueCache.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/devices.h"
 
 namespace cms::alpakatools {
