+++ HeterogeneousCore/AlpakaCore/src/alpaka/EDMetadataSentry.cc
@@ -1,7 +1,7 @@
-#include "HeterogeneousCore/AlpakaCore/interface/alpaka/chooseDevice.h"
 #include "HeterogeneousCore/AlpakaCore/interface/alpaka/EDMetadataSentry.h"
-#include "HeterogeneousCore/AlpakaCore/interface/EventCache.h"
-#include "HeterogeneousCore/AlpakaCore/interface/QueueCache.h"
+#include "HeterogeneousCore/AlpakaCore/interface/alpaka/chooseDevice.h"
+#include "HeterogeneousCore/AlpakaInterface/interface/EventCache.h"
+#include "HeterogeneousCore/AlpakaInterface/interface/QueueCache.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
   namespace detail {
